this commit addresses acceptance criteria items:
5) Display in a table below, iSites that have already begun to be imported. Show: iSite keyword, iSIte course title, and iSite course term
6) Text in the blue box is TBD, and will be provided.
7) There is a "more" button on the right in case the table listing the former iSites is longer than 3 rows.

I added hidden fields to the form to pass in the course title and term. These are then added to the details of the async job which in turn are able to be viewed in the output table on the bottom of the page.

Jira ticket# https://jira.huit.harvard.edu/browse/TLT-2098